+++ backend/run_tasks.py
@@ -6,6 +6,8 @@ import os
 from time import time, sleep
 from common import log, LOG_SQLALCHEMY_RT
 log.init('rt', log_sqlalchemy=LOG_SQLALCHEMY_RT)
+if __name__ == "__main__":
+    log.info('Starting run_tasks.py...')
 from common.sql_db_sync import wait_for_database
 if bool(os.getenv("WAIT_FOR_PG_AT_START", "")):
     wait_for_database()