+++ frontend/src/pages/QueryDocuments/models/docTasksQuery.ts
@@ -1,5 +1,8 @@
+import type { DocTasksGVDBsCfgState } from "../stores/useDocTasksGVDBsCfg";
+
 export type DocTasksQuery = {
   gvdbs_id: number;
+  gvdbs_cfg_json?: DocTasksGVDBsCfgState;
   gllms_id: number;
   gc_id: number;
   short_name: string;