+++ frontend/src/pages/QueryDocuments/stores/useDocTaskOptionsStore.ts
@@ -1,4 +1,5 @@
 import { createResettableStore } from "../../../api/createResettableStore";
+import type { SearchKwargsType, SearchType } from "./useDocTasksGVDBsCfg";
 
 export type GroupContexts = {
   gc_id: number;
@@ -22,10 +23,16 @@ export type GroupVDBs = {
   gvdbs_status: "success" | "warning" | "danger";
 };
 
+export type GVDBsCfgDefaults = {
+  search_type: SearchType;
+  search_kwargs_per_type: Record<SearchType, SearchKwargsType>;
+};
+
 export interface DocTaskOptionsResponse {
   group_contexts: GroupContexts[];
   group_llms: GroupLLMs[];
   group_vdbs: GroupVDBs[];
+  gvdbs_cfg_defaults: GVDBsCfgDefaults;
 }
 
 interface DocTaskOptionsState {
@@ -37,7 +44,19 @@ interface DocTaskOptionsState {
 
 export const useDocTaskOptionsStore =
   createResettableStore<DocTaskOptionsState>((set) => ({
-    data: { group_contexts: [], group_llms: [], group_vdbs: [] },
+    data: {
+      group_contexts: [],
+      group_llms: [],
+      group_vdbs: [],
+      gvdbs_cfg_defaults: {
+        search_type: "similarity",
+        search_kwargs_per_type: {
+          similarity: { k: 10 },
+          mmr: { k: 10, fetch_k: 20, lambda_mult: 0.5 },
+          similarity_score_threshold: { k: 10, score_threshold: 0.5 },
+        },
+      },
+    },
     setData: (data) => set({ data }),
     needReload: true,
     setNeedReload: (needReload: boolean) => set({ needReload }),